Use vim input mode in ZSH
This is, of course, a work in progress. I'm not happy with the key
bindings yet, but my intention is to add or fix them when I encounter
problems rather than trying to do everything up front.

I have copied some bindings I found earlier though, but that's just to
have something to read to understand how one does bindings in the first
place.