Put swap files in their original locations
I'm just so tired of naming conflicts and bogus "ATTENTION:"s. The swap
files are gitignored by default right now so I'll never commit one. It's
also hidden so it should seldom get in the way.
Only thing is, a crashed vim will leave a lot of files behind, which in
turn will be part of archives made for that directory. That sucks, of
course.