Use diff-so-fancy more focused
Replacing the entire pager was a pain as I don't want diff-so-fancy in
git log, for example.

I also removed the -F option from less and the search pattern bit as it
was just causing pains instead of being nice. Added -i and -n options to
less too.

Use more compact diffs from Git 2.9 and use diff-so-fancy in interactive
prompts.