Improve install flow for AUR
Now each package is installed one at a time, so if many new packages are
pending and one of them fails to compile, then the rest of the packages
are still compiled and installed.

Next time you re-run the command only the failing packages from last
time will be compiled and installed.