taskwarrior: Low priority *decreases* urgency
Turns out the coefficient is not a coefficient on the urgency, but a
coefficient that is multiplied by the "field". So for tags or priority,
that number is always 1.0.

Then all these multipliers are summed for the final urgency score.

Adding a coefficient of 0.5 does not halve the score, it adds another
0.5 points. Adding a coefficient of 2.0 does not double the score; it
usually just adds 2.0 points.

The field number is not 1.0 in cases like "age", "annotations", etc.