Keyboard: Simplify JSON; overlay supports size
The JSON got simplified by not relying on "a" (alignment) anymore. The
online editor adds these in order to make the keycap text simpler, but
this makes all logic and handling so much harder. By manually editing
the JSON now I get to choose how to deal with font sizes, alignment,
etc. and then I can use the same method on *all* keys which makes
everything so much simpler.

The overlay system can now change the size of text on keys.