Major change: Remove Hyper key and switch to Super
I've reprogrammed my keyboard to move some keys around, which makes it
easier to use Super over my Hyper key.

I decided to go back to something simpler because of the following
reasons (in no particular order):

1. This is simpler to reason about because most software understands
   what Super is, but might not understand what my Hyper is.

2. "Hyper" as a name is used by some esoteric keyboards to mean "All
   modifiers at the same time", which is not what this is.

3. I was not using Super anyway; why leave unused modifiers when they
   could be merged together?

4. Now I get to have access to Caps Lock while I didn't before (adding
   another modifier requires sacrificing a key after all).
   I mean, why shouldn't I get to scream online? IT'S NOT FAIR!!!!

5. This makes it possible to use my machine even if Xmodmap is reset
   (which happens when keyboards are re-plugged). E.g. I could use my
   config in a pinch where I cannot add my Hyper key.

Let's see how this works out. The downside is that if I'm forced to use
another keyboard (like a laptop's built-in keyboard) that I cannot
reprogram, then the Super key might be in a really bad spot and suddenly
I cannot use that keyboard comfortably.
This could perhaps be fixed with an additional Xmodmap file that remaps
Caps Lock to Super, and then run it on-demand. Let's see how I'll deal
with it the next time I'm using the laptop's keyboard (I suspect this
Tuesday).