Use Ruby-native ripper-tags over ctags
This greatly improves matches found in Ruby files. For example, even
with the custom regex patterns in the ctags config, this line:
  attr_reader :foo, :bar, :baz
still only emits `:baz` as these regexes can only match a single match
at a time.

In addition, this case:
  class Foo
    BAR = "bar"
  end
would emit the root constant `BAR` with the custom regex, but with
ripper-tags the `BAR` constant is correctly nested under the `Foo`
parent.

…and this is just the tip of the iceberg. I've left the custom ctags
settings in case ripper-tags isn't available in some contexts, or if an
older version of the hooks is running somewhere.

@teoljungberg, you might find this interesting too.

The gem is installed under all Ruby versions in RVM and added to the
global gemset so it is automatically installed when a new version of
Ruby is installed.