Add backup scripts (replace deja-dup)
deja-dup is pretty much impossible to get to work outside of GNOME. I've
been trying for months, but it just never wants to auto-activate.

deja-dup is mostly a GNOME frontend to duplicity, so let's switch to
using duplicity manually instead. On a positive note, this also means
that I get to keep my excludes file inside this repo. Very nice!