Arch: Install copy (not symlink) of user-dirs.dirs
The file was constantly regenerated from some sort of background job,
which caused dotfiles installation to error out a lot.

It's probably best to just create a copy of it and install a new version
when it has been updated.

Also creates all the directories specified in the file (which is the
main cause for the generated changes to happen; non-existent directories
were removed).