i3: Make autorandr work without preset
The "reflow screeens" shortcut it meant to fix screens in exotic
configurations too, so that they try to pick the "best config". If there
is no set config, fall back to setting up the screens in a horizontal
config.

User can then change it around a bit and then save the configuration.