Arch: Download updates in the background
This only downloads the updates in the background, but it means that
when I trigger an update I don't have to wait for the download to
finish as well. Should make it quicker to throw in a quick upgrade
before rebooting, or before logging in when a machine is booted.