Polybar: Change looks / colors / different bars
Different bars between primary screen and secondary screen. This allows
me to force tray to always be on primary screen, and to remove some
redundant information that don't belong on both screens.

Added a slight background on the bar and removed most gray colors to fit
better with my new wallpaper.