Get rid of two tray icons: Bluetooth and network
Trying to reduce the number of tray icons. Wifi is already covered by
the polybar entry and tydra's wifi menu shortcut.

Bluetooth is a bit harder. Added a tydra item to open the blueman
manager, but to quickly connect to my headphones is still something I'd
like a shortcut for (just like one for toggling audio output).