parcellite: Don't trim newlines; I misunderstood
I thought this was "trim" as in "Remove on the edges", like PHP's trim
function, or other languages' strip() function/method. However, this was
actually "Replace all", so if I copied three lines of something, then
all newlines were replaced with a space and put on the same line.

Not at all what I wanted.