Start working on my own rubocop.yml file
I'm adding rules as I go. It's going to take a while before it matches
my personal preferences, at which point I should also probably sort it
and explain the rules a bit.

But for now, this is enough to get rubocop off my back when working on
my own scripts.