Vim: Remove relativenumber
I don't like it. I prefer being able to look at stack traces on the
right and then look back to the left and see the line number. I also
like the other way around, when trying to write a command or comment
that refers to a line number on the right, I prefer being able to glance
to the side to see the line number.