Switch screen locking solution
Since some update, light-locker has stopped working correctly. Sometimes
the LightDM screen is never initialized, which requires me to sign in
via a TTY, extract ENV variables from a running X process, then call
`dm-tool send-to-greeter` to be able to sign in. Not very nice.

The added benefit of this new solution is that I don't get the constant
screen flickering and unlocking is now much faster.

The locking-monitoring scripts became a lot simpler now and that whole
tool can be removed. Now music will auto-resume if it was playing at the
time of the screen lock.