Improve lock-screen command when playerctl is slow
playerctl is very slow when using spotifyd sometimes, which makes the
lock screen take almost a minute to appear.

Now lock the screen first, then look at playerctl's current status,
which places that long delay behind the lock screen. In addition, place
a short timeout in front of playerctl so it can not take longer than a
few seconds.