Vim: [experiment] Free <tab> and <s-tab>
I use them all the time, every day. It's part of my muscle memory.
However, I wonder if it needs to be, or if placing them under <leader>
would make them much slower. This is an experiment to see if I can
unlock <tab> and <s-tab> for other things that LSP provides, or other
cool plugins, etc.

If I don't like this I will return.