Awesome: Feature to focus one window temporarily
Like "I need to focus on this particular window right now". I usually
use fullscreen for this, but maybe this would be better as the bar is
still visible. Fullscreen could be the last "ulra-super-deluxe focus"
where not even the bar can distract me anymore.

In either case, it's more like an experiement.