Awesome: Add window-selector and minimize actions
This is really nice for windows I don't really care about right now;
e.g. an application server running in development mode. I could always
jump to the window if I need it, but it generally does not need to take
up space in the layout that are better served for other windows.