Disable kglobalaccel as best as I can
Pretty much all KDE apps requires the kglobalaccel service and will
start it. By default it steals a lot of bindings that I want to use by
other apps, and it only happens after I've run a KDE program that starts
this service.

There's no reliable way to disable this service, but I *could* produce
a minimal config with no binding configured and keep that instead.
I hope this works.