Vim: Remove custom json formatting options
This mucks up Ruby buffers as the nested heredoc support now loads the
JSON filetype. These options were recommended by the old JSON.vim
plugin, but that is no longer used and I cannot see that there are any
other recommendations inside the plugin that replaced it.

Better to start with vanilla and add back whatever I find to be useful
later.