Neovim: Rebuild my entire config
This new config is much more Lua-based than before. Most things are now
in Lua instead of Vimscript.

Re-evaluated all of my plugins and mappings to take advantage of
Neovim's latest features, like tree sitter and built-in LSP.

ALE and CoC are both gone, and a lot of plugins have been replaced with
more modern alternatives written in Lua instead.

The config is still not perfect; after all it took years to get the
previous config semi-stable, so this will also take some time to settle,
but it's now a productive daily driver for me.