Remove whole email setup
I'm tired of it. It never gets good and I constantly avoid using it and
have been for a long time now. Let's just get rid of it completely.

It's hard to beat the built-in webmail clients, honestly. Perhaps some
more modern email client could do it, but then I usually miss all the
GMail features not in IMAP (Snooze, tags, archive, etc.).

Even if I manage to get all of this to a nice working state, I don't
want to have to spend time maintaining it either.