Add bin to find an icon from an app name
This should be useful in a lot of scripting. The icon lookup logic is
very complicated, and this is not fully spec compliant since we do not
even try to care about icon size and just look up using some simple
logic.

Things not considered:

1. Theme is assumed to only inherit at most one theme at a time.
2. Themes are assumed to eventually explicitly inherit "hicolor" instead
   of manually searching that theme again.
3. Size constraints are not considered at all.
4. Almost nothing in the theme config file is considered.

However, we do correctly look in all the spec-compliant places in the
correct order. That should count for something, right?