Awesome: Rework window focusing shortcuts
A lot of things had to move around to get this to work.

HJKL now focuses windows by direction (globally), and ; and ' (right of
HJKL) focuses screens left and right.

N and P is now for "next" and "previous" (by index), in case that makes
more sense in some layouts.

Minimize and unminimize is moved into the Chord bindings.