Neovim: Avoid double diagnostics in Ruby
Solargraph is usually configured to use Solargraph in the repos I'm
working on currently. It sucks that it will default to Rubocop on
unconfigured projects, but right now it's very fiddly to try to set up
different LSP servers in different projects, so I'm just going to set
my global config to what works currently.

I think this will be changed around for me later on to work better in
the general case.