Add an example catppuccin gradient wallpaper
It looks like crap since it's using a weird color depth and also isn't
adjusted for my screen size. It's just an example to test how the
contrast looks like with a gradient background.

If I want to commit to this I will do so by generating it using Cairo
inside of Awesome WM instead. It doesn't look extremely likely right
now, though.