Awesome: Rethink module reloading setup
This new setup should be more "vanilla"-like in that we use normal
requires everywhere instead of a custom function.
In addition, modules can register cleanup functions even when not being
required with a special function.

The special function that runs the initialization function should now
only be called from rc.lua or from within another initialization
function. This removes the confusion on when to use the special require
and when to use a normal one. It's also safe to use the initializer on a
module that does not need it.

I hope this will be a lot easier to work with.

The biggest change in approach is to not strive for a fully reloadable
environment for each and every file. Instead, I want to be able to
reload whatever I have under configuration/ or global modules like
daemons and similar.

Individual widgets are easiest to just do a normal reload on and then
rebuilding the entire tree, or perhaps even just restarting Awesome.