Neovim: Update vim-cmp and config
The max_item_count option has been removed, so I had to fix priority
settings in a better way instead.

Note that sources get an implied priority as the reverse index in the
list, so the last source gets `1` and the first gets `n`. Here I want to
boost the last one even further down.