Experiment: Add ags
Let's see if this is something that is nice to work with or not. Right
now it seems like:

  * Type support is broken.
  * Documentation is fairly sparse.
  * Hyprland support is basic.

It might not be worth it for me right now; it'll require a lot of time
investment to get working the way I want it to.
However, when it actually works it seems to be really awesome!

I should consider it later.