Install Ghostty
I love Wezterm, and I think a lot of important features are missing from
Ghostty. But on the other hand Wezterm has had a lot of issues the last
year or so, and development have ground to a halt.

After testing it for a bit as my main terminal, it's missing too many
features and is too slow for me. But I can keep it around for the fun
"In-game terminal" shader stuff and other fun nonsense.

It's also good as a backup terminal in case something happens, or if I
want to test a TUI/CLI in another popular terminal to see if it works or
not.